extract Debug stuff
