+++ main/src/ca/uwaterloo/flix/language/phase/constraintgeneration/TypingConstraint.scala
@@ -16,46 +16,33 @@
 package ca.uwaterloo.flix.language.phase.constraintgeneration
 
 import ca.uwaterloo.flix.language.ast.{Kind, Level, SourceLocation, Symbol, Type}
-import ca.uwaterloo.flix.language.phase.unification.Substitution
 
 
 sealed trait TypingConstraint {
 
   lazy val index: (Int, Int, Int) = this match {
-    case TypingConstraint.Equality(tvar1: Type.Var, Type.Pure, _) => (0, 0, 0)
-    case TypingConstraint.Equality(Type.Pure, tvar2: Type.Var, _) => (0, 0, 0)
+    case TypingConstraint.Equality(_: Type.Var, Type.Pure, _) => (0, 0, 0)
+    case TypingConstraint.Equality(Type.Pure, _: Type.Var, _) => (0, 0, 0)
     case TypingConstraint.Equality(tvar1: Type.Var, tvar2: Type.Var, _) if tvar1 != tvar2 => (0, 0, 0)
-    case TypingConstraint.Purification(sym, eff1, eff2, level, prov, nested) => (0, 0, 0)
-    case TypingConstraint.Equality(tpe1, tpe2, prov) =>
-      val tvars = (tpe1.typeVars ++ tpe2.typeVars)
+    case TypingConstraint.Purification(_, _, _, _, _, _) => (0, 0, 0)
+    case TypingConstraint.Equality(tpe1, tpe2, _) =>
+      val tvars = tpe1.typeVars ++ tpe2.typeVars
       val effTvars = tvars.filter(_.kind == Kind.Eff)
       (1, effTvars.size, tvars.size)
-    case TypingConstraint.Class(sym, tpe, loc) => (2, 0, 0)
+    case TypingConstraint.Class(_, _, _) => (2, 0, 0)
   }
 
   override def toString: String = this match {
-    case TypingConstraint.Equality(tpe1, tpe2, prov) => s"$tpe1 ~ $tpe2"
-    case TypingConstraint.Class(sym, tpe, loc) => s"$sym[$tpe]"
-    case TypingConstraint.Purification(sym, eff1, eff2, level, prov, nested) => s"$eff1 ~ ($eff2)[$sym ↦ Pure] ∧ $nested"
+    case TypingConstraint.Equality(tpe1, tpe2, _) => s"$tpe1 ~ $tpe2"
+    case TypingConstraint.Class(sym, tpe, _) => s"$sym[$tpe]"
+    case TypingConstraint.Purification(sym, eff1, eff2, _, _, nested) => s"$eff1 ~ ($eff2)[$sym ↦ Pure] ∧ $nested"
   }
 
   def numVars: Int = this match {
-    case TypingConstraint.Equality(tpe1, tpe2, prov) => tpe1.typeVars.size + tpe2.typeVars.size
-    case TypingConstraint.Class(sym, tpe, lc) => tpe.typeVars.size
-    case TypingConstraint.Purification(sym, eff1, eff2, level, prov, nested) => eff1.typeVars.size + eff2.typeVars.size
+    case TypingConstraint.Equality(tpe1, tpe2, _) => tpe1.typeVars.size + tpe2.typeVars.size
+    case TypingConstraint.Class(_, tpe, _) => tpe.typeVars.size
+    case TypingConstraint.Purification(_, eff1, eff2, _, _, _) => eff1.typeVars.size + eff2.typeVars.size
   }
-
-  private def toSubDot: String = this match {
-    case TypingConstraint.Equality(tpe1, tpe2, prov) => s"""$dotId [label = "$tpe1 ~ $tpe2"];"""
-    case TypingConstraint.Class(sym, tpe, loc) => s"""$dotId [label = "$sym[$tpe]"];"""
-    case TypingConstraint.Purification(sym, eff1, eff2, level, prov, nested) =>
-      val header = s"""$dotId [label = "$eff1 ~ ($eff2)[$sym ↦ Pure]"];"""
-      val children = nested.map(_.toSubDot)
-      val edges = nested.map { child => s"$dotId -> ${child.dotId};" }
-      (header :: children ::: edges).mkString("\n")
-  }
-
-  private def dotId: Int = System.identityHashCode(this)
 
   def loc: SourceLocation
 }
@@ -74,63 +61,6 @@ object TypingConstraint {
   case class Purification(sym: Symbol.KindedTypeVarSym, eff1: Type, eff2: Type, level: Level, prov: Provenance, nested: List[TypingConstraint]) extends TypingConstraint {
     def loc = prov.loc
   }
-
-  def toDot(constrs: List[TypingConstraint]): String = {
-    val contents = constrs.map(_.toSubDot)
-    val edges = constrs.map { constr => s"root -> ${constr.dotId};" }
-    format((
-      "digraph Constraints {" ::
-        "rankdir = \"LR\";" ::
-        "node [shape=\"box\"];" ::
-        contents :::
-        edges :::
-        List("}")
-      ).mkString("\n"))
-  }
-
-  def toDotWithSubst(constrs: List[TypingConstraint], subst: Substitution): String = {
-    val contents = constrs.map(_.toSubDot)
-    val edges = constrs.map { constr => s"root -> ${constr.dotId};" }
-    val substPart = subst.m.toList.sortBy(_._1).flatMap {
-      case (sym, tpe) =>
-        val tvar = Type.Var(sym, SourceLocation.Unknown)
-        val from = System.nanoTime()
-        val to = System.nanoTime()
-        List(
-          s"""$from [label = "$tvar"];""",
-          s"""$to [label = "$tpe"];""",
-          s"$from -> $to;"
-        )
-    }
-    val substSubgraph =
-      "subgraph cluster_Subst {" ::
-        "label = \"Substitution\";" ::
-        "style = \"filled\";" ::
-        "invisL [style=\"invis\"];" :: // these invisible boxes help align the substitution
-        "invisR [style=\"invis\"];" ::
-        "invisL -> invisR [style=\"invis\"];" ::
-        substPart :::
-        List("}")
-
-    val constrsSubgraph =
-      "subgraph Constraints {" ::
-        "root" ::
-        contents :::
-        edges :::
-        List("}")
-
-    format((
-      "digraph Constraints {" ::
-        "rankdir = \"LR\";" ::
-        "node [shape=\"box\"];" ::
-        substSubgraph :::
-        constrsSubgraph :::
-        "invisR -> root [style=\"invis\"];" ::
-        List("}")
-      ).mkString("\n"))
-  }
-
-  private def format(s: String): String = s.replace("\\", "\\\\")
 
   sealed trait Provenance {
     def loc: SourceLocation
